feat: refactor overview page components and replace deprecated charts with new implementations
